[FIX] Change demo data field, display_name for name. [MOV] Rename company and partner py files [ADD] Add ResPartnerStateField Class to new py file
