+++ partner_state/demo/res.partner.state_field.csv
@@ -1,3 +1,3 @@
 id,field_id/id,track,approval
 state_field_1,base.field_res_partner_active,True,True
-state_field_2,base.field_res_partner_display_name,True,True
+state_field_2,base.field_res_partner_name,True,True
