+++ partner_state/models/__init__.py
@@ -2,5 +2,6 @@
 # For copyright and license notices, see __manifest__.py file in module root
 # directory
 ##############################################################################
-from . import company
-from . import partner
+from . import res_company
+from . import res_partner
+from . import res_partner_state_field
